rename: assets > static -> constants 변경 (#19)
