+++ src/components/composites/login/social/SocialLogin.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import { useToast } from '@/hooks';
-import { TOAST_MESSAGE } from '@/assets';
 import KakaoSymbolIcon from '@/assets/icon/ic_kakaoSymbol.svg?react';
 import NaverSymbolIcon from '@/assets/icon/ic_naverSymbol.svg?react';
+import { TOAST_MESSAGE } from '@/constants';
 import * as S from './SocialLogin.styled';
 
 const SocialLogin = () => {
