+++ src/components/composites/main/popularBook/desktop/PopularBookDesktop.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 
 import { SegmentedButton } from '@/components';
-import { POPULAR_BOOK_DATE_FILTER_OPTIONS } from '@/assets';
+import { POPULAR_BOOK_DATE_FILTER_OPTIONS } from '@/constants';
 import type { PopularBookCardType, SelectOptionType } from '@/types';
 import * as S from './PopularBookDesktop.styled';
